// src/components/SideBarMenu/enum.ts
export enum FAQ_FLOW {
  STEP_PHONE = "authentication-phone",
  STEP_EMAIL = "authentication-email",
  STEP_VERIFY_PHONE = "verify-number",
  STEP_SELECT_DOCUMENT = "select-document",
  STEP_CAPTURE = "capture-documents",
  STEP_SELFIE = "selfie",
  STEP_INFO = "personal-info",
  STEP_FRONT = "front",
  STEP_BACK = "back",
  STEP_INTRO = "intro",
  STEP_PRODUCT_SELECTION = "business-or-personal",
  STEP_DEPOSIT_SUB_PRODUCTS = "deposit-sub-products",
}